requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.5.0
opencv-python>=4.7.0
python-chess>=1.9.0
tqdm>=4.65.0s